The "contrast" setting is actually a float within the range 0.5-2.0 which needs to converted to a fixed point number for the driver

+++ Framework/Commons/Source/Representations/Infrastructure/CameraSettings.h
@@ -82,7 +82,7 @@ namespace naoth
     bool backlightCompensation;
     int brightness;
     int cameraSelection;
-    int contrast;
+    double contrast;
     int exposure;
     bool fadeToBlack;
     int gain;
